DTSPB-1739 - Prev Identified execs and any other partners (#1487)
* add previously identified executors to title and clearing page

* only show prev identified execs if we have them for solicitor

* Add isPractitionerAnExecutor field

* Fix typo and fix exec names bug

* Fix previously identified executors issue for not named as exec

* fix bug in removing sol from applying list

* remove isProbatePractitioner

* attend e2e for prev named execs

* add any other partners applying

* e2e for not named, applying

* exec names bugfix

* make any other ptnrs always show

* fux unit test issues and resolve exec list mapper bug

* add shouldTransformForDeceasedDetails test

* only show any other partners when we need to

* amend e2es for any other partners

+++ src/main/java/uk/gov/hmcts/probate/transformer/solicitorexecutors/ExecutorsTransformer.java
@@ -86,6 +86,9 @@ public class ExecutorsTransformer {
         List<CollectionMember<AdditionalExecutorNotApplying>> execsNotApplying =
                 createCaseworkerNotApplyingList(caseData);
 
+        execsApplying = setExecutorApplyingListWithSolicitorInfo(execsApplying, caseData);
+        execsNotApplying = setExecutorNotApplyingListWithSolicitorInfo(execsNotApplying, caseData);
+
         // Format exec lists into strings
         String execsApplyingNames = FormattingService.createExecsApplyingNames(execsApplying);
         String execsNotApplyingNames = FormattingService.createExecsNotApplyingNames(execsNotApplying);
@@ -93,6 +96,10 @@ public class ExecutorsTransformer {
         // Set builder with exec strings
         builder.solsIdentifiedApplyingExecs(execsApplyingNames);
         builder.solsIdentifiedNotApplyingExecs(execsNotApplyingNames);
+
+        // Set the copies of this data needed to get round ccd limitations
+        builder.solsIdentifiedApplyingExecsCcdCopy(execsApplyingNames);
+        builder.solsIdentifiedNotApplyingExecsCcdCopy(execsNotApplyingNames);
     }
 
     public List<CollectionMember<AdditionalExecutorApplying>> createCaseworkerApplyingList(CaseData caseData) {
@@ -167,7 +174,7 @@ public class ExecutorsTransformer {
             // Add solicitor to applying list
             execsApplying = executorListMapperService.addSolicitorToApplyingList(caseData, execsApplying);
 
-        } else if (!isSolicitorExecutor(caseData) || !isSolicitorApplying(caseData)) {
+        } else {
 
             // Remove solicitor from applying executor list
             execsApplying = executorListMapperService.removeSolicitorFromApplyingList(execsApplying);
@@ -180,12 +187,12 @@ public class ExecutorsTransformer {
             List<CollectionMember<AdditionalExecutorNotApplying>> execsNotApplying, CaseData caseData) {
 
         // Transform list
-        if (!isSolicitorApplying(caseData)) {
+        if (!isSolicitorApplying(caseData) && isSolicitorNamedInWillAsAnExecutor(caseData)) {
 
             // Add solicitor to not applying list
             execsNotApplying = executorListMapperService.addSolicitorToNotApplyingList(caseData, execsNotApplying);
 
-        } else if (isSolicitorApplying(caseData)) {
+        } else {
 
             // Remove solicitor from not applying executor list
             execsNotApplying = executorListMapperService.removeSolicitorFromNotApplyingList(execsNotApplying);
@@ -217,14 +224,13 @@ public class ExecutorsTransformer {
         caseData.setSolsPrimaryExecutorNotApplyingReason(null);
     }
 
-    public void setFieldsIfSolicitorIsNotExecutor(CaseData caseData) {
-        if (!isSolicitorExecutor(caseData)) {
-            caseData.setSolsSolicitorIsApplying(NO);
+    public void setFieldsIfSolicitorIsNotNamedInWillAsAnExecutor(CaseData caseData) {
+        if (!isSolicitorNamedInWillAsAnExecutor(caseData)) {
             caseData.setSolsSolicitorNotApplyingReason(null);
         }
     }
 
-    protected boolean isSolicitorExecutor(CaseData caseData) {
+    protected boolean isSolicitorNamedInWillAsAnExecutor(CaseData caseData) {
         return YES.equals(caseData.getSolsSolicitorIsExec());
     }
